try-fix

+++ src/main/java/io/github/sefiraat/networks/network/NetworkRoot.java
@@ -211,19 +211,19 @@ public class NetworkRoot extends NetworkNode {
     }
 
     @Nonnull
-    public Map<ItemStack, Integer> getAllNetworkItems() {
-        final Map<ItemStack, Integer> itemStacks = new HashMap<>();
+    public Map<ItemStack, Long> getAllNetworkItems() {
+        final Map<ItemStack, Long> itemStacks = new HashMap<>();
 
         // Barrels
         for (BarrelIdentity barrelIdentity : getBarrels()) {
-            final Integer currentAmount = itemStacks.get(barrelIdentity.getItemStack());
-            final int newAmount;
+            final Long currentAmount = itemStacks.get(barrelIdentity.getItemStack());
+            final long newAmount;
             if (currentAmount == null) {
                 newAmount = barrelIdentity.getAmount();
             } else {
                 long newLong = (long) currentAmount + (long) barrelIdentity.getAmount();
-                if (newLong > Integer.MAX_VALUE) {
-                    newAmount = Integer.MAX_VALUE;
+                if (newLong > Long.MAX_VALUE) {
+                    newAmount = Long.MAX_VALUE;
                 } else {
                     newAmount = currentAmount + barrelIdentity.getAmount();
                 }
@@ -237,14 +237,14 @@ public class NetworkRoot extends NetworkNode {
                 continue;
             }
             final ItemStack clone = StackUtils.getAsQuantity(itemStack, 1);
-            final Integer currentAmount = itemStacks.get(clone);
-            final int newAmount;
+            final Long currentAmount = itemStacks.get(clone);
+            final long newAmount;
             if (currentAmount == null) {
                 newAmount = itemStack.getAmount();
             } else {
                 long newLong = (long) currentAmount + (long) itemStack.getAmount();
-                if (newLong > Integer.MAX_VALUE) {
-                    newAmount = Integer.MAX_VALUE;
+                if (newLong > Long.MAX_VALUE) {
+                    newAmount = Long.MAX_VALUE;
                 } else {
                     newAmount = currentAmount + itemStack.getAmount();
                 }
@@ -260,14 +260,14 @@ public class NetworkRoot extends NetworkNode {
                     continue;
                 }
                 final ItemStack clone = StackUtils.getAsQuantity(itemStack, 1);
-                final Integer currentAmount = itemStacks.get(clone);
-                final int newAmount;
+                final Long currentAmount = itemStacks.get(clone);
+                final long newAmount;
                 if (currentAmount == null) {
                     newAmount = itemStack.getAmount();
                 } else {
                     long newLong = (long) currentAmount + (long) itemStack.getAmount();
-                    if (newLong > Integer.MAX_VALUE) {
-                        newAmount = Integer.MAX_VALUE;
+                    if (newLong > Long.MAX_VALUE) {
+                        newAmount = Long.MAX_VALUE;
                     } else {
                         newAmount = currentAmount + itemStack.getAmount();
                     }
@@ -283,15 +283,15 @@ public class NetworkRoot extends NetworkNode {
 
                     clone.setAmount(1);
 
-                    final Integer currentAmount = itemStacks.get(clone);
-                    int newAmount;
+                    final Long currentAmount = itemStacks.get(clone);
+                    long newAmount;
 
                     if (currentAmount == null) {
                         newAmount = itemStack.getAmount();
                     } else {
                         long newLong = (long) currentAmount + (long) itemStack.getAmount();
-                        if (newLong > Integer.MAX_VALUE) {
-                            newAmount = Integer.MAX_VALUE;
+                        if (newLong > Long.MAX_VALUE) {
+                            newAmount = Long.MAX_VALUE;
                         } else {
                             newAmount = currentAmount + itemStack.getAmount();
                         }
